Add example of module reference problems
handle_ref=1 means that the module increases a reference to itself until
the delayed work finished. When handle_ref=0, the user is able to rmmod
the module, and as soon as the delayed work is ran, there is a crash
since the function that is being called is not in memory anymore.